requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
streamlit==1.44.1
python-multipart==0.0.20
sse-starlette==2.2.1
pydantic==2.11.3
pydantic-settings==2.8.1
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.14
langchain-groq==0.3.2
langchain-huggingface==0.1.2
langchain-chroma==0.2.2
langchain-pinecone==0.2.5
langchain-qdrant==0.2.0
chromadb==1.0.5
pypdf==5.4.0
sentence-transformers==4.1.0
rank-bm25==0.2.2
httpx==0.28.1
python-dotenv==1.1.0
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
qdrant-client==1.13.3
pinecone==6.0.2
